Split into modules; rename and repackage as shapeless3
