--- core/src/main/scala/shapeless/test/package.scala
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2014-19 Miles Sabin
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package shapeless
-
-import scala.quoted._
-
-package object test {
-  def typed[T](t : => T): Unit = {}
-
-  def sameTyped[T](t1: => T)(t2: => T): Unit = {}
-
-  inline def showType[T]: String = ${ impl[T] }
-
-  inline def showType[T](t: => T): String = ${ impl[T] }
-
-  def impl[T: Type](using qctx: QuoteContext): Expr[String] = {
-    import qctx.tasty._
-    Expr(typeOf[T].show)
-  }
-}
