--- core/src/main/scala/shapeless/test/typechecking.scala
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2013-19 Miles Sabin
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package shapeless
-package test
-
-import scala.compiletime.testing._
-
-/**
- * A utility which ensures that a code fragment does not typecheck.
- */
-object illTyped {
-  inline def apply(code: String): Unit = assert(!typeChecks(code))
-  inline def apply(code: String, expected: String): Unit = apply(code)
-}
