+++ modules/data/src/main/scala/shapeless3/data/hlists.scala
@@ -1,0 +1,63 @@
+/*
+ * Copyright (c) 2011-19 Miles Sabin
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package shapeless3.data
+
+/**
+ * `HList` ADT base trait.
+ *
+ * @author Miles Sabin
+ */
+sealed trait HList extends Product with Serializable
+
+/**
+ * Non-empty `HList` element type.
+ *
+ * @author Miles Sabin
+ */
+final case class :*:[+H, +T <: HList](head : H, tail : T) extends HList {
+  override def toString = head match {
+    case _: :*:[_, _] => s"($head) :: $tail"
+    case _ => s"$head :*: $tail"
+  }
+}
+
+/**
+ * Empty `HList` element type.
+ *
+ * @author Miles Sabin
+ */
+sealed trait HNil extends HList {
+  def :*:[H](h : H): H :*: HNil = shapeless3.data.:*:(h, this)
+  override def toString = "HNil"
+}
+
+object HList {
+  implicit class HListOps[L <: HList](l : L) {
+    /**
+     * Prepend the argument element to this `HList`.
+     */
+    def :*:[H](h : H) : H :*: L = shapeless3.data.:*:(h, l)
+  }
+}
+
+/**
+ * Empty `HList` value.
+ *
+ * @author Miles Sabin
+ */
+
+case object HNil extends HNil
